implemented the new features

+++ admin/class-wp-list-table/wt-appointments-list.php
@@ -13,10 +13,18 @@ class appointments_List_Table extends WP_List_Table{
     }
 
     function column_name($item) {
-        $actions = array(
-            'View' => sprintf('<a href="?page=report-incident&view=%s">%s</a>', $item['cb'], __('View', 'what')),
-            'delete' => sprintf('<a href="?page=%s&action=delete&id=%s">%s</a>', $_REQUEST['page'], $item['cb'], __('Delete', 'what')),
-        );
+    	if( isset( $_REQUEST['ap-type'] ) && in_array($_REQUEST['ap-type'], array('past','cancelled')) ){
+    		$actions = array(
+            	//'View' => sprintf('<a href="?page=report-incident&view=%s">%s</a>', $item['cb'], __('View', 'what')),
+	            'delete' => sprintf('<a href="?page=%s&action=delete&id=%s">%s</a>', $_REQUEST['page'], $item['cb'], __('Delete', 'what')),
+	        );
+    	}
+    	else{
+    		$actions = array(
+	            'cancel' => sprintf('<a href="?page=%s&action=cancel&id=%s">%s</a>', $_REQUEST['page'], $item['cb'], __('Cancel', 'what')),
+	        );
+    	}
+        
 
        return sprintf('%s %s',
             $item['name'],
@@ -29,6 +37,8 @@ class appointments_List_Table extends WP_List_Table{
 			case 'cb':
 			case 'name':
 			case 'date':
+				return $item[ $column_name ];
+			case 'date_user':
 				return $item[ $column_name ];
 			case 'app_email':
 				return $item[ $column_name ];
@@ -46,6 +56,7 @@ class appointments_List_Table extends WP_List_Table{
 			'cb' => '<input type="checkbox" />', //Render a checkbox instead of text
 			'name' => 'Name',
 			'date' => "Appointments's Time",
+			'date_user' => "User's Time",
 			'app_email' => 'Email',
 			'app_booked_time' => 'Booked Date',
 			'app_comments' => 'Comments',
@@ -87,9 +98,16 @@ class appointments_List_Table extends WP_List_Table{
     }
 	
 	function get_bulk_actions() {
-		$actions = array(
-			'delete' => 'Delete',
-	  	);
+		if( isset( $_REQUEST['ap-type'] ) && in_array($_REQUEST['ap-type'], array('past','cancelled')) ){
+			$actions = array(
+				'delete' => 'Delete',
+		  	);
+		}
+		else{
+			$actions = array(
+				'cancel' => 'Cancel',
+		  	);
+		}
 	  	return $actions;
 	} 
 			
@@ -102,8 +120,36 @@ class appointments_List_Table extends WP_List_Table{
             if (is_array( $ids )) $ids = implode(',', $ids);
 			
 			if (!empty( $ids )) {
-                $wpdb->query("DELETE FROM $table_appointments WHERE id IN($ids)");
+                $wpdb->query("DELETE FROM $table_appointments WHERE app_id IN($ids)");
 			}
+        }
+        else if ('cancel' === $this->current_action()) {
+        	$ids = isset($_REQUEST['id']) ? $_REQUEST['id'] : array();
+        	if (!empty( $ids )) {
+        		if( is_array( $ids ) ){
+	        		foreach ($ids as $key => $id) {
+		        		$status = $wpdb->update( 
+				            $table_appointments, 
+				            array('app_status' => 'cancelled'),
+				            array( 'app_id' => $id  ) 
+				        );
+				        if($status !== false){
+				        	$this->send_user_cancelation_email( $id );
+				        }
+	        		}        			
+        		}
+        		else{
+        			$status = $wpdb->update( 
+			            $table_appointments, 
+			            array('app_status' => 'cancelled'),
+			            array( 'app_id' => $ids  ) 
+			        );
+
+			        if($status !== false){
+			        	$this->send_user_cancelation_email( $ids );
+			        }
+        		}
+        	}
         }
     }
 	
@@ -120,12 +166,15 @@ class appointments_List_Table extends WP_List_Table{
 		$appointments = '';
 		$ap_type = ( isset( $_REQUEST['ap-type'] ) ) ? $_REQUEST['ap-type'] : '';
 		if( !empty( $ap_type ) && $ap_type == 'past' ){
-			$query = "SELECT * FROM {$table_appointments} AS wr_p WHERE wr_p.app_time <= '$date' order by app_id desc";
+			$query = "SELECT * FROM {$table_appointments} AS wr_p WHERE wr_p.app_time <= '$date' AND wr_p.app_status = 'upcoming' order by app_id desc";
+		}
+		else if( !empty( $ap_type ) && $ap_type == 'cancelled' ){
+			$query = "SELECT * FROM {$table_appointments} AS wr_p WHERE wr_p.app_status = 'cancelled' order by app_id desc";
 		}
 		else{
-			$query = "SELECT * FROM {$table_appointments} AS wr_p WHERE wr_p.app_time >= '$date' order by app_id desc";
+			$query = "SELECT * FROM {$table_appointments} AS wr_p WHERE wr_p.app_time >= '$date' AND wr_p.app_status = 'upcoming' order by app_id desc";
 		}
-		
+
 		$appointments = $wpdb->get_results( $query );
 			
 		if(!empty( $appointments )) {
@@ -136,11 +185,13 @@ class appointments_List_Table extends WP_List_Table{
             	$app_email = $value->app_email;
             	$app_comments = $value->app_comments;
             	$app_booked_time = $value->app_booked_time;
-				
+            	$date_user = $value->app_user_time . "<br>".$value->app_timezone;
+
 				$data[] = array(
-					'cb' => $value->id,
+					'cb' => $value->app_id,
 					'name' => $name,					
 					'date' => $date,					
+					'date_user' => $date_user,					
 					'app_email' 	=> $app_email,
 					'app_comments' 	=> $app_comments,
 					'app_booked_time' 	=> $app_booked_time,
@@ -150,6 +201,7 @@ class appointments_List_Table extends WP_List_Table{
 		return $data;
 	} 
 
+	
 	function get_views(){
 		   $views = array();
 		   $current = ( !empty($_REQUEST['ap-type']) ? $_REQUEST['ap-type'] : 'upcoming');
@@ -164,7 +216,54 @@ class appointments_List_Table extends WP_List_Table{
 		   $class = ($current == 'past' ? ' class="current"' :'');
 		   $views['past'] = "<a href='{$past_url}' {$class} >Past</a>";
 
+		   $cancelled_url = add_query_arg('ap-type','cancelled');
+		   $class = ($current == 'cancelled' ? ' class="current"' :'');
+		   $views['cancelled'] = "<a href='{$cancelled_url}' {$class} >Cancelled</a>";
+
 		   return $views;
+	}
+
+	/**
+	 * Functionality for send cancelation email to user
+	 *
+	 * @since    1.0.0
+	 * @access   private
+	 */
+	private function send_user_cancelation_email( $id ) {
+
+		global $wpdb;
+        $table_appointments = $wpdb->prefix . "wt_appointments";
+
+		$query = "SELECT * FROM {$table_appointments} AS wr_p WHERE wr_p.app_id IN($id) ";
+	    $results = $wpdb->get_results( $query );
+
+	    if(  !empty( $results )){
+
+		    $wt_appointments = get_option('wt_appointments_options');
+		    $email = new Wt_Appointment_Email( $this->plugin_name, $this->version );
+		    
+		    foreach ($results as $key => $value) {
+
+				$request = array();
+		        $request['subject'] = $wt_appointments['cancel']['subject'];
+			    $request['heading'] = $wt_appointments['cancel']['heading'];
+			    $request['body'] = apply_filters('the_content', get_option('wt_cancel_body'));
+
+			    $event_date = date_i18n( "M d, Y", strtotime( $value->app_user_time ) );
+			    $event_time = date_i18n( "h:i A", strtotime( $value->app_user_time ) );
+
+			    $request['sent_to'] = $value->app_email;
+			    $request['event_date'] = $event_date;
+			    $request['event_time'] = $event_time;
+			    $request['invitee_full_name'] = $value->app_name;
+			    $request['event_host_name'] = "Admin";
+			    $request['save_file'] = true;
+
+				$email->load_edm_template("", "", $request);
+
+		    }
 		}
+
+	}
 }
 ?>
